+++ packages/daemon/test/service/job-runner.test.ts
@@ -1,12 +1,8 @@
-import * as ChileProcess from 'child_process';
 import * as sinon from 'sinon';
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as core from '@pipcook/pipcook-core';
-import { PluginPackage } from '@pipcook/costa';
 import { strict as assert } from 'assert';
-import { JobEntity } from '../../src/model/job';
-import { PipelineEntity } from '../../src/model/pipeline';
 import { JobRunner } from '../../src/service/pipeline';
 import { JobStatusChangeEvent } from '../../src/service/trace-manager';
 
@@ -22,40 +18,6 @@ const runner = new JobRunner({
 describe('test JobRunner', () => {
   afterEach(() => {
     sinon.restore();
-  });
-
-  it('#generate output', async () => {
-    // @ts-ignore
-    sinon.replace(ChileProcess, 'exec', (cmd, opts, cb) => {
-      assert.equal(cmd, 'npm init -y', 'exec command check');
-      assert.deepEqual(opts, { cwd: '/home/output' }, 'exec option check');
-      cb();
-    });
-    const mockFsRemove = sinon.stub(fs, 'remove').resolves(true);
-    const mockFsEnsureDir = sinon.stub(fs, 'ensureDir').resolves(true);
-    const mockFsReadJson = sinon.stub(fs, 'readJSON').resolves({});
-    const mockFsCopy = sinon.stub(fs, 'copy').resolves(true);
-    const mockFsOutputJson = sinon.stub(fs, 'outputJSON').resolves({});
-    const mockFsCompressTarFile = sinon.stub(core, 'compressTarFile').resolves();
-
-    await runner.generateOutput({ id: 'mockId' } as JobEntity, {
-      modelPath: 'mockPath',
-      modelPlugin: { name: 'modelPlugin', version: 'mockVersion' } as PluginPackage,
-      pipeline: {} as PipelineEntity,
-      dataProcess: {
-        version: 'mockVersion',
-        name: 'dataProcess'
-      } as PluginPackage,
-      workingDir: '/home',
-      template: 'mock template'
-    });
-    // @ts-ignore
-    assert.ok(mockFsRemove.calledOnceWith('/home/output'), 'check mockFsRemove');
-    assert.ok(mockFsEnsureDir.calledOnceWith('/home/output'), 'check mockFsEnsureDir');
-    assert.ok(mockFsReadJson.called, 'check mockFsReadJson');
-    assert.ok(mockFsCopy.called, 'check mockFsCopy');
-    assert.ok(mockFsOutputJson.called, 'check mockFsOutputJson');
-    assert.ok(mockFsCompressTarFile.calledOnceWith('/home/output', '/home/output.tar.gz'), 'check mockFsCompressTarFile');
   });
 
   it('#test getParam with param string and empty extra field', async () => {
@@ -99,31 +61,31 @@ describe('test JobRunner', () => {
       datasetRoot: __dirname
     });
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('dataCollect');
+      runner.assertPlugin('dataCollect');
     }, 'dataCollect check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('dataAccess');
+      runner.assertPlugin('dataAccess');
     }, 'dataAccess check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('dataProcess');
+      runner.assertPlugin('dataProcess');
     }, 'dataProcess check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('datasetProcess');
+      runner.assertPlugin('datasetProcess');
     }, 'datasetProcess check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('modelLoad');
+      runner.assertPlugin('modelLoad');
      }, 'modelLoad check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('modelDefine');
+      runner.assertPlugin('modelDefine');
     }, 'modelDefine check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('modelTrain');
+      runner.assertPlugin('modelTrain');
     }, 'modelTrain check');
     assert.doesNotThrow(() => {
-      runner.verifyPlugin('modelEvaluate');
+      runner.assertPlugin('modelEvaluate');
     }, 'modelEvaluate check');
     assert.throws(() => {
-      runner.verifyPlugin('unkwon plugin type');
+      runner.assertPlugin('unkwon plugin type');
     }, 'unkwon plugin type');
   });
