+++ packages/daemon/test/service/pipeline.test.ts
@@ -1,5 +1,6 @@
 import * as sinon from 'sinon';
 import * as fs from 'fs-extra';
+import * as ChileProcess from 'child_process';
 import * as core from '@pipcook/pipcook-core';
 import { PluginPackage } from '@pipcook/costa';
 import * as path from 'path';
@@ -145,7 +146,40 @@ describe('test the pipeline service', () => {
     // TODO(Feely): check args
     assert.ok(mockFsRemove.calledTwice, 'check mockFsRemove');
   });
+  it('#generate output', async () => {
+    const pipeline: PipelineService = await app.applicationContext.getAsync<PipelineService>('pipelineService');
+    // @ts-ignore
+    sinon.replace(ChileProcess, 'exec', (cmd, opts, cb) => {
+      assert.equal(cmd, 'npm init -y', 'exec command check');
+      assert.deepEqual(opts, { cwd: '/home/output' }, 'exec option check');
+      cb();
+    });
+    const mockFsRemove = sinon.stub(fs, 'remove').resolves(true);
+    const mockFsEnsureDir = sinon.stub(fs, 'ensureDir').resolves(true);
+    const mockFsReadJson = sinon.stub(fs, 'readJSON').resolves({});
+    const mockFsCopy = sinon.stub(fs, 'copy').resolves(true);
+    const mockFsOutputJson = sinon.stub(fs, 'outputJSON').resolves({ });
+    const mockFsCompressTarFile = sinon.stub(core, 'compressTarFile').resolves();
 
+    await pipeline.generateOutput({ id: 'mockId' } as JobEntity, {
+      modelPath: 'mockPath',
+      modelPlugin: { name: 'modelPlugin', version: 'mockVersion' } as PluginPackage,
+      pipeline: {} as PipelineEntity,
+      dataProcess: {
+        version: 'mockVersion',
+        name: 'dataProcess'
+      } as PluginPackage,
+      workingDir: '/home',
+      template: 'mock template'
+    });
+    // @ts-ignore
+    assert.ok(mockFsRemove.calledOnceWith('/home/output'), 'check mockFsRemove');
+    assert.ok(mockFsEnsureDir.calledOnceWith('/home/output'), 'check mockFsEnsureDir');
+    assert.ok(mockFsReadJson.called, 'check mockFsReadJson');
+    assert.ok(mockFsCopy.called, 'check mockFsCopy');
+    assert.ok(mockFsOutputJson.called, 'check mockFsOutputJson');
+    assert.ok(mockFsCompressTarFile.calledOnceWith('/home/output', '/home/output.tar.gz'), 'check mockFsCompressTarFile');
+  });
   it('#should get output tar by job id', async () => {
     const pipeline: PipelineService = await app.applicationContext.getAsync<PipelineService>('pipelineService');
     assert.equal(pipeline.getOutputTarByJobId('mockId'), path.join(core.constants.PIPCOOK_RUN, 'mockId', 'output.tar.gz'));
